refactor(sdk): move the whisper verbs onto rooms
A whisper is a room concept — XEP-0045 §7.5, a private message to one
occupant, addressed `room/nick` with the `muc#user` marker so the room never
sees it. It lived on the messaging module only because that is where message
sending happened to be written.

`sendWhisper` and `sendWhisperChatState` move to `rooms`, unchanged. They
never used the messaging module's internals: both build a stanza from
`deps.sendStanza`, `deps.stores` and `deps.emitSDK`, which every module has.

Receiving stays where it is. One stanza router claims `<message>`, and an
incoming whisper travels the same decrypt, fastening and correction path as
any other room message; splitting that would mean duplicating it. The test
file follows the send side and is renamed, since its subject was already
`MUC Whispers`.
