+++ packages/fluux-sdk/src/core/modules/MUC.ts
@@ -7,6 +7,7 @@ import { generateUUID } from '../../utils/uuid'
 import { generateQuickChatSlug } from '../wordlist'
 import { hasStableOccupantIdentity, isNonAnonymousRoom, isPrivateRoom } from '../roomCapabilities'
 import { parseBookmarkItem } from '../bookmarkItem'
+import { createOriginIdElement } from './messagingUtils'
 import { PRESENCE_PRIORITY } from '../config'
 import {
   NS_MUC,
@@ -28,6 +29,8 @@ import {
   NS_COMMANDS,
   NS_RETRACT,
   NS_MESSAGE_MODERATE,
+  NS_CHATSTATES,
+  NS_HINTS,
 } from '../namespaces'
 import type {
   Room,
@@ -40,6 +43,8 @@ import type {
   RSMResponse,
   AdminRoom,
   RoomFeatures,
+  RoomMessage,
+  ChatStateNotification,
 } from '../types'
 import type { StoredRoomMessage } from '../types/message-internal'
 import { parseXMPPError, formatXMPPError, hasErrorCondition } from '../../utils/xmppError'
@@ -1547,19 +1552,75 @@ export class MUC extends BaseModule {
   }
 
   /**
-   * Set the room subject/topic (XEP-0045).
+   * XEP-0045 §7.5: send a private message ("whisper") to a single room
+   * occupant. Unlike `messages.sendMessage`, this preserves the `/nick`
+   * resource (sending to a person strips it) and emits `room:whisper`
+   * instead of `chat:message` so the message is tracked as a private room
+   * message rather than a 1:1 conversation. The XEP-0334 `<no-store>` hint
+   * keeps it off the server archive; it is still persisted locally.
    *
-   * Sends a groupchat message with a `<subject>` element.
-   * The server will echo this back as a room subject change.
-   *
-   * @param roomJid - The room JID
-   * @param subject - The new subject text
-   *
-   * @example
-   * ```typescript
-   * await client.rooms.setSubject('room@conference.example.com', 'New topic for discussion')
-   * ```
+   * @param roomJid bare room JID, e.g. 'room@conference.example.com'
+   * @param nick    target occupant's nickname
+   * @param body    message text
+   * @returns the generated message id
    */
+  async sendWhisper(roomJid: string, nick: string, body: string): Promise<string> {
+    const id = generateUUID()
+    const to = `${roomJid}/${nick}`
+
+    const message = xml('message', { to, type: 'chat', id },
+      xml('body', {}, body),
+      xml('active', { xmlns: NS_CHATSTATES }),
+      xml('x', { xmlns: NS_MUC_USER }),
+      createOriginIdElement(id),
+      xml('no-store', { xmlns: NS_HINTS }),
+    )
+    await this.deps.sendStanza(message)
+
+    const room = this.deps.stores?.room.getRoom(roomJid)
+    if (!room) logWarn(`sendWhisper: room ${roomJid} not found in store — sender nick will be empty`)
+    const ourNick = room?.nickname || ''
+    // Counterpart's stable occupant-id (XEP-0421), resolved from the live occupant
+    // list, so a persisted thread can be re-bound to the same person later.
+    const targetOccupantId = room?.occupants.get(nick)?.occupantId
+    const whisper: RoomMessage = {
+      type: 'groupchat',
+      id,
+      originId: id,
+      roomJid,
+      from: `${roomJid}/${ourNick}`,
+      nick: ourNick,
+      body,
+      timestamp: new Date(),
+      isOutgoing: true,
+      isPrivate: true,
+      whisperWith: nick,
+      ...(targetOccupantId && { whisperWithOccupantId: targetOccupantId }),
+    }
+    this.deps.emitSDK('room:whisper', {
+      roomJid,
+      message: whisper,
+      incrementUnread: false,
+      incrementMentions: false,
+    })
+    return id
+  }
+
+  /**
+   * Send a chat state (XEP-0085) privately to a single room occupant — the typing
+   * indicator for a whisper (XEP-0045 §7.5). Unlike sendChatState on the room,
+   * which broadcasts to the room, this addresses room/nick with the muc#user marker
+   * and a no-store hint, so the room never sees that you are whispering.
+   */
+  async sendWhisperChatState(roomJid: string, nick: string, state: ChatStateNotification): Promise<void> {
+    const message = xml('message', { to: `${roomJid}/${nick}`, type: 'chat' },
+      xml(state, { xmlns: NS_CHATSTATES }),
+      xml('x', { xmlns: NS_MUC_USER }),
+      xml('no-store', { xmlns: NS_HINTS }),
+    )
+    await this.deps.sendStanza(message)
+  }
+
   /**
    * Re-fetch recent history for every joined room that supports archiving.
    *
@@ -1578,6 +1639,20 @@ export class MUC extends BaseModule {
     await this.mam.forceCatchUpAllRooms(options)
   }
 
+  /**
+   * Set the room subject/topic (XEP-0045).
+   *
+   * Sends a groupchat message with a `<subject>` element.
+   * The server will echo this back as a room subject change.
+   *
+   * @param roomJid - The room JID
+   * @param subject - The new subject text
+   *
+   * @example
+   * ```typescript
+   * await client.rooms.setSubject('room@conference.example.com', 'New topic for discussion')
+   * ```
+   */
   async setSubject(roomJid: string, subject: string): Promise<void> {
     const msg = xml(
       'message',
